requirements.txt
Flask==2.2.3
google-generativeai==0.0.5  # Replace with the correct version if known
requests==2.28.1
